requirements.txt
Flask==2.3.2
requests==2.28.1
python-dotenv==0.21.1
gunicorn==20.1.0
